+++ lib/main.dart
@@ -1,13 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_blue/flutter_blue.dart';
-import 'package:dropdown_button2/dropdown_button2.dart';
-// import 'package:flutter_localizations/flutter_localizations.dart';
 
-// import 'app_localizations.dart';
+import 'ParametresPage.dart';
+import 'SauvegardePage.dart';
+import 'TelechargementPage.dart';
 
-// import 'package:flutter_gen/gen_l10n/app_localizations.dart';
-// translation : avec fichier .arb
-// https://medium.com/@pinkesh.earth/easy-way-to-internationalize-and-localize-your-flutter-app-flutterbeads-e250fbcf63e2
 void main() {
   runApp(MyApp());
 }
@@ -161,15 +158,6 @@ class _MyHomePageState extends State<MyHomePage> {
     );
   }
 
-  // Widget switch(ScanResult r){
-
-  //   return Switch(
-  //                 activeColor: Colors.green,
-  //                 value: isSwitched,
-  //                 onChanged: (change) {
-  //                   print("Clic");
-  //                 });  }
-
   void onTap(ScanResult r) {
     print('${r.device.name}');
   }
@@ -226,904 +214,8 @@ class _MyHomePageState extends State<MyHomePage> {
 }
 
 // La page Paramètres
-class ParametresPage extends StatefulWidget {
-  const ParametresPage({super.key});
-
-  @override
-  State<ParametresPage> createState() => _ParametresPageState();
-}
-
-class _ParametresPageState extends State<ParametresPage> {
-  final List<String> items = [
-    'Français',
-    'Anglais',
-  ];
-  final List<String> temp = [
-    'Celcius (°C)',
-    'Fahrenheit (°F)',
-  ];
-  String? selectedValue;
-  String? selectedTemp;
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        // centerTitle: true,
-        title: const Text(
-          // AppLocalizations.of(context)?.translate('first_string'),
-          "Paramètres",
-          textAlign: TextAlign.left,
-          selectionColor: Color.fromARGB(255, 0, 0, 0),
-        ),
-      ),
-      drawer: drawer(context),
-      body: Container(
-        child: Column(
-          mainAxisSize: MainAxisSize.max,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Padding(
-              padding: const EdgeInsets.all(5),
-            ),
-            Padding(
-              padding: const EdgeInsets.all(10),
-              child: GestureDetector(
-                onTap: () {
-                  Navigator.pop(context);
-                  // Navigator.push(context,
-                  //     MaterialPageRoute(builder: (context) => MyApp()));
-                },
-                child: const Text.rich(
-                  TextSpan(
-                    children: <InlineSpan>[
-                      WidgetSpan(
-                          child: Icon(
-                        Icons.arrow_back_ios,
-                        color: Color.fromARGB(255, 0, 0, 0),
-                        size: 18,
-                      )),
-                      TextSpan(text: 'Retour'),
-                    ],
-                  ),
-                  textAlign: TextAlign.center,
-                  style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.all(10),
-              child: Image.asset('assets/images/logo.png'),
-            ),
-            Padding(
-                padding: const EdgeInsets.all(10),
-                child: Container(
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(50),
-                    gradient: LinearGradient(
-                      begin: new Alignment(-1, -1),
-                      end: new Alignment(-0.5, 1),
-                      colors: [
-                        Color.fromARGB(255, 255, 249, 191),
-                        Color.fromARGB(255, 255, 251, 214),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8)
-                      ],
-                      // center: Alignment.topLeft,
-                      // radius: 0.8,
-                    ),
-                  ),
-                  child: ListTile(
-                    tileColor: const Color.fromARGB(33, 33, 149, 243),
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(50)),
-                    title: Text(
-                      "Panel Solaire Connect v1.0",
-                      style: TextStyle(fontSize: 20),
-                    ),
-                    subtitle: Text("Copyright ©"),
-                  ),
-                )),
-            Padding(
-                padding: const EdgeInsets.all(10),
-                child: Row(
-                  mainAxisSize: MainAxisSize.max,
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Padding(
-                      padding: const EdgeInsets.all(10),
-                      child: Text(
-                        "Langue",
-                        style: TextStyle(
-                            fontSize: 18, fontWeight: FontWeight.bold),
-                      ),
-                    ),
-                    DropdownButtonHideUnderline(
-                      child: DropdownButton2(
-                        hint: Text(
-                          'Sélectionner',
-                          style: TextStyle(
-                            fontSize: 18,
-                            color: Theme.of(context).hintColor,
-                          ),
-                        ),
-                        items: items
-                            .map((item) => DropdownMenuItem<String>(
-                                  value: item,
-                                  child: Text(
-                                    item,
-                                    style: const TextStyle(
-                                      fontSize: 18,
-                                    ),
-                                  ),
-                                ))
-                            .toList(),
-                        value: selectedValue,
-                        onChanged: (value) {
-                          setState(() {
-                            selectedValue = value as String;
-                          });
-                          return langue(selectedValue);
-                        },
-                        buttonHeight: 40,
-                        buttonWidth: 140,
-                        itemHeight: 40,
-                      ),
-                    ),
-                  ],
-                )),
-            Padding(
-                padding: const EdgeInsets.all(10),
-                child: Row(
-                  mainAxisSize: MainAxisSize.max,
-                  mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                  children: [
-                    Padding(
-                      padding: const EdgeInsets.all(10),
-                      child: Text(
-                        "Unité de température",
-                        style: TextStyle(
-                            fontSize: 18, fontWeight: FontWeight.bold),
-                      ),
-                    ),
-                    DropdownButtonHideUnderline(
-                      child: DropdownButton2(
-                        isExpanded: true,
-                        hint: Text(
-                          'Sélectionner',
-                          style: TextStyle(
-                            fontSize: 18,
-                            color: Theme.of(context).hintColor,
-                          ),
-                        ),
-                        items: temp
-                            .map((temp) => DropdownMenuItem<String>(
-                                  value: temp,
-                                  child: Text(
-                                    temp,
-                                    style: const TextStyle(
-                                      fontSize: 18,
-                                    ),
-                                  ),
-                                ))
-                            .toList(),
-                        value: selectedTemp,
-                        onChanged: (value) {
-                          setState(() {
-                            selectedTemp = value as String;
-                          });
-                          return langue(selectedTemp);
-                        },
-                        buttonHeight: 40,
-                        buttonWidth: 160,
-                        itemHeight: 40,
-                        dropdownWidth: 160,
-                      ),
-                    ),
-                  ],
-                )),
-          ],
-        ),
-      ),
-    );
-  }
-
-  void langue(select) {
-    print(select);
-  }
-
-  void temperature(select) {
-    print(select);
-  }
-}
 
 // La page Sauvegarde de fichiers
-
-class SauvegardePage extends StatefulWidget {
-  const SauvegardePage({super.key});
-
-  @override
-  State<SauvegardePage> createState() => _SauvegardePageState();
-}
-
-class _SauvegardePageState extends State<SauvegardePage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        // centerTitle: true,
-        title: const Text(
-          "Sauvegarde de fichiers",
-          textAlign: TextAlign.left,
-        ),
-      ),
-      drawer: Drawer(child: drawer(context)),
-      body: Container(
-        child: Column(
-          mainAxisSize: MainAxisSize.max,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Padding(
-              padding: const EdgeInsets.all(5),
-            ),
-            // retour
-            Padding(
-              padding: const EdgeInsets.all(10),
-              child: GestureDetector(
-                onTap: () {
-                  Navigator.pop(context);
-                  // Navigator.push(context,
-                  //     MaterialPageRoute(builder: (context) => MyApp()));
-                },
-                child: const Text.rich(
-                  TextSpan(
-                    children: <InlineSpan>[
-                      WidgetSpan(
-                          child: Icon(
-                        Icons.arrow_back_ios,
-                        color: Color.fromARGB(255, 0, 0, 0),
-                        size: 18,
-                      )),
-                      TextSpan(text: 'Retour'),
-                    ],
-                  ),
-                  textAlign: TextAlign.center,
-                  style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.all(5),
-            ),
-            Padding(
-                padding: const EdgeInsets.all(8),
-                child: Container(
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(50),
-                    gradient: LinearGradient(
-                      begin: new Alignment(-1, -1),
-                      end: new Alignment(-0.5, 1),
-                      colors: [
-                        Color.fromARGB(255, 255, 249, 191),
-                        Color.fromARGB(255, 255, 251, 214),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8)
-                      ],
-                      // center: Alignment.topLeft,
-                      // radius: 0.8,
-                    ),
-                  ),
-                  child: ListTile(
-                    tileColor: const Color.fromARGB(33, 33, 149, 243),
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(50)),
-                    title: Text(
-                      "Smart Lithium (100Ah)",
-                      style: TextStyle(fontSize: 20),
-                    ),
-                    subtitle: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text("1856233252663"),
-                        Text.rich(
-                          TextSpan(
-                            children: <InlineSpan>[
-                              WidgetSpan(
-                                  child: Icon(
-                                Icons.arrow_forward_ios,
-                                color: Color.fromARGB(255, 0, 0, 0),
-                                size: 18,
-                              )),
-                              TextSpan(
-                                  text: '15-09-2022 09:46',
-                                  style: TextStyle(
-                                      fontSize: 14,
-                                      fontWeight: FontWeight.bold)),
-                            ],
-                          ),
-                          textAlign: TextAlign.center,
-                          style: TextStyle(fontSize: 20),
-                        )
-                      ],
-                    ),
-                    leading: Wrap(
-                      spacing: 12, // space between two icons
-                      children: <Widget>[
-                        // Icon(Icons.wifi_2_bar), // icon-1
-                        Icon(
-                          Icons.battery_charging_full,
-                          size: 36.0,
-                        ), // icon-2
-                      ],
-                    ),
-                    trailing: Wrap(
-                      spacing: 12, // space between two icons
-                      children: <Widget>[
-                        // icon-1
-                        Icon(
-                          Icons.file_download,
-                          size: 36.0,
-                        ), // icon-2
-                      ],
-                    ),
-                  ),
-                )),
-            Padding(
-                padding: const EdgeInsets.all(8),
-                child: Container(
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(50),
-                    gradient: LinearGradient(
-                      begin: new Alignment(-1, -1),
-                      end: new Alignment(-0.5, 1),
-                      colors: [
-                        Color.fromARGB(255, 255, 249, 191),
-                        Color.fromARGB(255, 255, 251, 214),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8)
-                      ],
-                      // center: Alignment.topLeft,
-                      // radius: 0.8,
-                    ),
-                  ),
-                  child: ListTile(
-                    leading: Wrap(
-                      spacing: 12, // space between two icons
-                      children: <Widget>[
-                        // Icon(Icons.wifi_2_bar), // icon-1
-                        Icon(
-                          Icons.battery_charging_full,
-                          size: 36.0,
-                        ), // icon-2
-                      ],
-                    ),
-                    trailing: Wrap(
-                      spacing: 12, // space between two icons
-                      children: <Widget>[
-                        // icon-1
-                        Icon(
-                          Icons.file_download,
-                          size: 36.0,
-                        ), // icon-2
-                      ],
-                    ),
-                    tileColor: const Color.fromARGB(33, 33, 149, 243),
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(50)),
-                    title: Text(
-                      "Smart Lithium (100Ah)",
-                      style: TextStyle(fontSize: 20),
-                    ),
-                    subtitle: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text("1856233285256"),
-                        Text.rich(
-                          TextSpan(
-                            children: <InlineSpan>[
-                              WidgetSpan(
-                                  child: Icon(
-                                Icons.arrow_forward_ios,
-                                color: Color.fromARGB(255, 0, 0, 0),
-                                size: 18,
-                              )),
-                              TextSpan(
-                                  text: '17-06-2022 04:52',
-                                  style: TextStyle(
-                                    fontSize: 14,
-                                    fontWeight: FontWeight.bold,
-                                  )),
-                            ],
-                          ),
-                          textAlign: TextAlign.center,
-                          style: TextStyle(fontSize: 20),
-                        )
-                      ],
-                    ),
-                  ),
-                )),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-// la page téléchargement
-class TelechargementPage extends StatefulWidget {
-  const TelechargementPage({super.key});
-
-  @override
-  State<TelechargementPage> createState() => _TelechargementPageState();
-}
-
-class _TelechargementPageState extends State<TelechargementPage> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      appBar: AppBar(
-        // centerTitle: true,
-        title: const Text(
-          "Téléchargements",
-          textAlign: TextAlign.left,
-        ),
-      ),
-      drawer: Drawer(child: drawer(context)),
-      body: Container(
-        child: Column(
-          mainAxisSize: MainAxisSize.max,
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            Padding(
-              padding: const EdgeInsets.all(5),
-            ),
-            // retour
-            Padding(
-              padding: const EdgeInsets.all(10),
-              child: GestureDetector(
-                onTap: () {
-                  Navigator.pop(context);
-                  // Navigator.push(context,
-                  //     MaterialPageRoute(builder: (context) => MyApp()));
-                },
-                child: const Text.rich(
-                  TextSpan(
-                    children: <InlineSpan>[
-                      WidgetSpan(
-                          child: Icon(
-                        Icons.arrow_back_ios,
-                        color: Color.fromARGB(255, 0, 0, 0),
-                        size: 18,
-                      )),
-                      TextSpan(text: 'Retour'),
-                    ],
-                  ),
-                  textAlign: TextAlign.center,
-                  style: TextStyle(fontSize: 16, fontWeight: FontWeight.bold),
-                ),
-              ),
-            ),
-            Padding(
-              padding: const EdgeInsets.all(5),
-            ),
-            Padding(
-                padding: const EdgeInsets.all(8),
-                child: Container(
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(50),
-                    gradient: LinearGradient(
-                      begin: new Alignment(-1, -1),
-                      end: new Alignment(-0.5, 1),
-                      colors: [
-                        Color.fromARGB(255, 255, 249, 191),
-                        Color.fromARGB(255, 255, 251, 214),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8)
-                      ],
-                      // center: Alignment.topLeft,
-                      // radius: 0.8,
-                    ),
-                  ),
-                  child: ListTile(
-                    tileColor: const Color.fromARGB(33, 33, 149, 243),
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(50)),
-                    title: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          "Batterie Lithiium LiFeP04 Smart",
-                          style: TextStyle(
-                            fontSize: 16,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                        Text.rich(
-                          TextSpan(
-                            children: <InlineSpan>[
-                              TextSpan(
-                                  text: '12,8V / 50-100-200 Ah',
-                                  style: TextStyle(
-                                    fontSize: 16,
-                                    fontWeight: FontWeight.bold,
-                                  )),
-                            ],
-                          ),
-                          textAlign: TextAlign.center,
-                          style: TextStyle(fontSize: 20),
-                        )
-                      ],
-                    ),
-                    subtitle: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          "Fiche Technique",
-                          style: TextStyle(
-                            fontSize: 16,
-                          ),
-                        ),
-                      ],
-                    ),
-                    leading: Wrap(
-                      spacing: 12, // space between two icons
-                      children: <Widget>[
-                        // Icon(Icons.wifi_2_bar), // icon-1
-                        Icon(
-                          Icons.battery_charging_full,
-                          size: 36.0,
-                        ), // icon-2
-                      ],
-                    ),
-                    trailing: Wrap(
-                      spacing: 12, // space between two icons
-                      children: <Widget>[
-                        // icon-1
-                        Icon(
-                          Icons.file_download,
-                          size: 36.0,
-                        ), // icon-2
-                      ],
-                    ),
-                  ),
-                )),
-            Padding(
-              padding: const EdgeInsets.all(8),
-              child: Container(
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(50),
-                    gradient: LinearGradient(
-                      begin: new Alignment(-1, -1),
-                      end: new Alignment(-0.5, 1),
-                      colors: [
-                        Color.fromARGB(255, 255, 249, 191),
-                        Color.fromARGB(255, 255, 251, 214),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8)
-                      ],
-                      // center: Alignment.topLeft,
-                      // radius: 0.8,
-                    ),
-                  ),
-                  child: ListTile(
-                    tileColor: const Color.fromARGB(33, 33, 149, 243),
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(50)),
-                    title: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          "Batterie Lithiium LiFeP04 Smart",
-                          style: TextStyle(
-                            fontSize: 16,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                        Text.rich(
-                          TextSpan(
-                            children: <InlineSpan>[
-                              TextSpan(
-                                  text: '12,8V /100 Ah (Varta)',
-                                  style: TextStyle(
-                                    fontSize: 16,
-                                    fontWeight: FontWeight.bold,
-                                  )),
-                            ],
-                          ),
-                          textAlign: TextAlign.center,
-                          style: TextStyle(fontSize: 20),
-                        )
-                      ],
-                    ),
-                    subtitle: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          "Fiche Technique",
-                          style: TextStyle(
-                            fontSize: 16,
-                          ),
-                        ),
-                      ],
-                    ),
-                    leading: Wrap(
-                      spacing: 12, // space between two icons
-                      children: <Widget>[
-                        // Icon(Icons.wifi_2_bar), // icon-1
-                        Icon(
-                          Icons.battery_charging_full,
-                          size: 36.0,
-                        ), // icon-2
-                      ],
-                    ),
-                    trailing: Wrap(
-                      spacing: 12, // space between two icons
-                      children: <Widget>[
-                        // icon-1
-                        Icon(
-                          Icons.file_download,
-                          size: 36.0,
-                        ), // icon-2
-                      ],
-                    ),
-                  )),
-            ),
-            Padding(
-              padding: const EdgeInsets.all(8),
-              child: Container(
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(50),
-                    gradient: LinearGradient(
-                      begin: new Alignment(-1, -1),
-                      end: new Alignment(-0.5, 1),
-                      colors: [
-                        Color.fromARGB(255, 255, 249, 191),
-                        Color.fromARGB(255, 255, 251, 214),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8)
-                      ],
-                      // center: Alignment.topLeft,
-                      // radius: 0.8,
-                    ),
-                  ),
-                  child: ListTile(
-                    tileColor: const Color.fromARGB(33, 33, 149, 243),
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(50)),
-                    title: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          "Batterie Lithiium LiFeP04 Smart",
-                          style: TextStyle(
-                            fontSize: 16,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                        Text.rich(
-                          TextSpan(
-                            children: <InlineSpan>[
-                              TextSpan(
-                                  text: '12,8V / 150 Ah (Varta)',
-                                  style: TextStyle(
-                                    fontSize: 16,
-                                    fontWeight: FontWeight.bold,
-                                  )),
-                            ],
-                          ),
-                          textAlign: TextAlign.center,
-                          style: TextStyle(fontSize: 20),
-                        )
-                      ],
-                    ),
-                    subtitle: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          "Fiche Technique",
-                          style: TextStyle(
-                            fontSize: 16,
-                          ),
-                        ),
-                      ],
-                    ),
-                    leading: Wrap(
-                      spacing: 12, // space between two icons
-                      children: <Widget>[
-                        // Icon(Icons.wifi_2_bar), // icon-1
-                        Icon(
-                          Icons.battery_charging_full,
-                          size: 36.0,
-                        ), // icon-2
-                      ],
-                    ),
-                    trailing: Wrap(
-                      spacing: 12, // space between two icons
-                      children: <Widget>[
-                        // icon-1
-                        Icon(
-                          Icons.file_download,
-                          size: 36.0,
-                        ), // icon-2
-                      ],
-                    ),
-                  )),
-            ),
-            Padding(
-                padding: const EdgeInsets.all(8),
-                child: Container(
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(50),
-                    gradient: LinearGradient(
-                      begin: new Alignment(-1, -1),
-                      end: new Alignment(-0.5, 1),
-                      colors: [
-                        Color.fromARGB(255, 255, 249, 191),
-                        Color.fromARGB(255, 255, 251, 214),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8)
-                      ],
-                      // center: Alignment.topLeft,
-                      // radius: 0.8,
-                    ),
-                  ),
-                  child: ListTile(
-                    tileColor: const Color.fromARGB(33, 33, 149, 243),
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(50)),
-                    title: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          "Batterie Lithiium LiFeP04 Smart",
-                          style: TextStyle(
-                            fontSize: 16,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                        Text.rich(
-                          TextSpan(
-                            children: <InlineSpan>[
-                              TextSpan(
-                                  text: '12,8V / 130 Ah (Sous siège)',
-                                  style: TextStyle(
-                                    fontSize: 16,
-                                    fontWeight: FontWeight.bold,
-                                  )),
-                            ],
-                          ),
-                          textAlign: TextAlign.center,
-                          style: TextStyle(fontSize: 20),
-                        )
-                      ],
-                    ),
-                    subtitle: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          "Fiche Technique",
-                          style: TextStyle(
-                            fontSize: 16,
-                          ),
-                        ),
-                      ],
-                    ),
-                    leading: Wrap(
-                      spacing: 12, // space between two icons
-                      children: <Widget>[
-                        // Icon(Icons.wifi_2_bar), // icon-1
-                        Icon(
-                          Icons.battery_charging_full,
-                          size: 36.0,
-                        ), // icon-2
-                      ],
-                    ),
-                    trailing: Wrap(
-                      spacing: 12, // space between two icons
-                      children: <Widget>[
-                        // icon-1
-                        Icon(
-                          Icons.file_download,
-                          size: 36.0,
-                        ), // icon-2
-                      ],
-                    ),
-                  ),
-                )),
-            Padding(
-              padding: const EdgeInsets.all(8),
-              child: Container(
-                  decoration: BoxDecoration(
-                    borderRadius: BorderRadius.circular(50),
-                    gradient: LinearGradient(
-                      begin: new Alignment(-1, -1),
-                      end: new Alignment(-0.5, 1),
-                      colors: [
-                        Color.fromARGB(255, 255, 249, 191),
-                        Color.fromARGB(255, 255, 251, 214),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8),
-                        Color(0xffe8e8e8)
-                      ],
-                      // center: Alignment.topLeft,
-                      // radius: 0.8,
-                    ),
-                  ),
-                  child: ListTile(
-                    tileColor: const Color.fromARGB(33, 33, 149, 243),
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(50)),
-                    title: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          "Batterie Lithiium LiFeP04 Smart",
-                          style: TextStyle(
-                            fontSize: 16,
-                            fontWeight: FontWeight.bold,
-                          ),
-                        ),
-                        Text.rich(
-                          TextSpan(
-                            children: <InlineSpan>[
-                              TextSpan(
-                                  text: '12,8V / 180 Ah (Sous siège)',
-                                  style: TextStyle(
-                                    fontSize: 16,
-                                    fontWeight: FontWeight.bold,
-                                  )),
-                            ],
-                          ),
-                          textAlign: TextAlign.center,
-                          style: TextStyle(fontSize: 20),
-                        )
-                      ],
-                    ),
-                    subtitle: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(
-                          "Fiche Technique",
-                          style: TextStyle(
-                            fontSize: 16,
-                          ),
-                        ),
-                      ],
-                    ),
-                    leading: Wrap(
-                      spacing: 12, // space between two icons
-                      children: <Widget>[
-                        // Icon(Icons.wifi_2_bar), // icon-1
-                        Icon(
-                          Icons.battery_charging_full,
-                          size: 36.0,
-                        ), // icon-2
-                      ],
-                    ),
-                    trailing: Wrap(
-                      spacing: 12, // space between two icons
-                      children: <Widget>[
-                        // icon-1
-                        Icon(
-                          Icons.file_download,
-                          size: 36.0,
-                        ), // icon-2
-                      ],
-                    ),
-                  )),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
 
 // mon drawer
